chore(ci): Make sure to test all platforms in CI (#32)
## Launch Checklist

<!-- Thanks for the PR! Feel free to add or remove items from the
checklist. -->

- [x] Confirm **your changes do not include backports from Mapbox
projects** (unless with compliant license) - if you are not sure about
this, please ask!

This PR makes sure that we have a test matrix of all supported targets
and exuecute it.
It is mostly to test where the issues are. Debugging CI locally is not
something that I think works.

Resolves #31

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Yuri Astrakhan <yuriastrakhan@gmail.com>