Document inclusion of the Maplibre Native license. (fixes #5) (#37)
This PR addresses licensing concerns in `README.md`.

As I understand the BSD 2-Clause, `maplibre-native-rs` is not required
to include the MapLibre Native license, since binary and source
artifacts are downloaded to the host machine during `cargo build`, and
thus are not "redistributed" by the crate as far as the license is
concerned.

On the other hand, any project building against this crate _would_ need
to include the license with any binary release, since these would
technically be "redistributing" the binary code against which they were
linked.

fixes #5

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>