chore(ci): add a step to require conventional pr titles (#40)
If we add release-plz, we likely want a new step in the CI that asserts
that PR-titles are formatted with [conventional commit
messages](https://flank.github.io/flank/pr_titles/).

I am not quite sure if the step gets executed correcty and if we can
assert on this. Only smething that I think can be checked once im main