chore: remove the option to `git checkout` to simplify our tool chain (#53)
currently, our toolchain depnds on git checkout for a few paths. This is
quite messy and if we would not have this component, things would be
much simpler. This path is also not asserted in CI.

This is a proposal to remove this component as it significantly impacts
how I can reason currently.
Since we use git, the code is not lost. If someone needs it, we can
still re-add it later.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>